requirements.txt
bokeh==3.1.0
catboost==1.2
jupyter==1.0.0
matplotlib==3.7.1
pandas==2.0.1
pip==23.1.2
# py-xgboost==1.7.4
# python==3.10.10
# pytorch-cuda==11.7
pytorch==2.0.0
pytorch-lightning==2.0.2
scikit-learn==1.2.2
scipy==1.8.1
sympy==1.11.1
tensorboard==2.13.0
interpret==0.3.2
julia==0.6.1
numpy==1.23.5
optuna==3.1.1
pysindy==1.7.5
pysr==0.13.0
tabulate==0.9.0
xgboost==1.7.5
lightgbm==3.3.5
lime==0.2.0.1
dice-ml==0.11
shap==0.45.0
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-cytoscape==1.0.0
dash-html-components==2.0.0
dash-table==5.0.0
tslearn==0.6.3
